Proof of concept for the flaw in formattedWidth of field.
The hypothesis is that 'textwidth' in MCField is actually incorrect
as it is the maximum absolute width of every MCLine. What is should
be is the maximum width of every MCLine *after* trailing breaks have
been removed.

This patch has a number of flaws - for example, the cursor will get
stuck at the end of lines - but does indicate that the problem is
as described.