[[ LCB ]] Improve builtin module emission and loading
This patch reworks the way builtin modules are handled.

When lc-compile is run in --outputc mode, it will now emit a C
byte array for each module which is passed on the command line
along with a unique global class instance which calls
MCScriptRegisterBuiltinModule in its constructor. This function takes
a fixed format static record which contains the module data and its
(C) initializer and finalizer. When called, the function will link
the static record into a singly-linked list of builtin modules.

When MCScriptInitialize is called, it will now iterate through this
list loading each module, and setting each one's initializer and
finalizer. It will then ensure each builtin module is usable.

The initializer and finalizer are now called by libscript itself -
the initializer when the module is made usable, and the finalizer
when the module is (finally) destroyed.

This new mechanism means that builtin modules do not all have to be
compiled together into one C file, but compiled and linked into the
engine separately.

Note: It is necessary to 'force_load' (Mac)  any libs containing
builtin modules when linking the final executable, otherwise the
global objects used for registration will be stripped.

Note: Currently the gyp files have only been updated for Mac - the
other platforms are still to be done.