[[ VS2015 ]] Add export definition mode to list_stub_symbols
This patch adds an extra parameter --exportdef to the
list_stub_symbols utility.

When this option is present, the utility will generate a Windows
export definition file suitable for use to build a shared library.

It is used to build revsecurity, by generating a def file which
contains only the exports required by the weak linking for
our uses of OpenSSL.