[[ Cleanup ]] Remove use of MCString from externalv1.cpp/.h
This patch removes the use of MCString from externalv1.cpp and
externalv1.h.

One use was to provide a 'CData' accessor for strings - i.e.
a pair (char*, uint32_t) representing a counted char buffer.
This has been replaced by an explicit MCExternalCData struct
type.

The other use was to provide keys as CData when iterating
through an array.